Combining if block the reach less indented code

+++ src/controllers/CBController.php
@@ -1021,10 +1021,9 @@ class CBController extends Controller {
 			$inputdata = Request::get($name);
 
 			//Insert Data Checkbox if Type Datatable
-			if($ro['type'] == 'checkbox') {
-				if($ro['relationship_table']) {
-					$datatable = explode(",",$ro['datatable'])[0];					
-					
+			if($ro['type'] == 'checkbox' && $ro['relationship_table']) {
+					$datatable = explode(",",$ro['datatable'])[0];
+
 					$foreignKey2 = CRUDBooster::getForeignKey($datatable,$ro['relationship_table']);
 					$foreignKey = CRUDBooster::getForeignKey($this->table,$ro['relationship_table']);
 					DB::table($ro['relationship_table'])->where($foreignKey,$id)->delete();
@@ -1039,16 +1038,12 @@ class CBController extends Controller {
 								]);
 						}
 					}
-					
-
-				}
 			}
 
 
-			if($ro['type'] == 'select2') {
-				if($ro['relationship_table']) {
-					$datatable = explode(",",$ro['datatable'])[0];					
-					
+			if($ro['type'] == 'select2' && $ro['relationship_table']) {
+					$datatable = explode(",",$ro['datatable'])[0];
+
 					$foreignKey2 = CRUDBooster::getForeignKey($datatable,$ro['relationship_table']);
 					$foreignKey = CRUDBooster::getForeignKey($this->table,$ro['relationship_table']);
 					DB::table($ro['relationship_table'])->where($foreignKey,$id)->delete();
@@ -1063,9 +1058,6 @@ class CBController extends Controller {
 								]);
 						}
 					}
-					
-
-				}
 			}
 
 			if($ro['type']=='child') {
